Rewrite launch screen using objects
I would like to integrate the launch screen into the app better, but it
is difficult to do.

I rewrote the launch screen to separate drawing from animation which
works better with shoes.  I created a Wave and Ship object which have
separate draw and animate methods.  The draw method puts things on the
screen and the animate moves it around based on the current frame.

This makes the launch_screen method much easier to understand as each
part (setup, draw, animate) is separate from the others.